Closes #12 by (hopefully) fixing my main annoyance indentation issue, the one where else would indent significantly to match the last if instead of the if I lined up with before I started typing.
Fixes the original issue problem (the pass issue) and the shift-o issue by fixing a bug, I got ssline using ssnum before setting ssnum. Once that was fixed then both issues started working as expected.

Lastly, the hash issue. I couldn't figure out a way to make this one work, I know the line it is failing on but it's such a niche case (an if statment that has a hash in a string) that I guess I will just deal with it when it happens because I couldn't figure out a true fix to the problem